Extract validation metrics to reduce file length
